realAPI branch: real api key received, integrated in configuration and service, some helpers models added. problem with fractional number (tempterature and wind speed) serializing

+++ WeatherWebAPI.Contracts/Models/WeatherResponse.cs
@@ -19,7 +19,7 @@ public class WeatherResponse
     /// <summary>
     /// Current temperature
     /// </summary>
-    public double Temperature { get; set; }
+    public decimal Temperature { get; set; }
 
     /// <summary>
     /// Weather description
@@ -34,5 +34,5 @@ public class WeatherResponse
     /// <summary>
     /// Speed of wind
     /// </summary>
-    public double WindSpeed { get; set; }
+    public float WindSpeed { get; set; }
 }
